Added shooting functionality and comments
Shooting is now possible. Left clicking will spawn a bullet at the
players location. It will rotate towards the mouse click and move
towards it at a constant rate. The rotation is a little messed up but I
don't want to take the time to make it perfect right now since the
bullet I'm using is not going to be the final asset and I don't want to
waste time just incase.